clean up user/role class
