+++ backend/src/main/java/com/bakdata/conquery/models/auth/subjects/PermissionOwner.java
@@ -1,36 +1,19 @@
 package com.bakdata.conquery.models.auth.subjects;
 
 import java.util.Arrays;
-import java.util.Collection;
-import java.util.EnumSet;
 import java.util.HashSet;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Optional;
 import java.util.Set;
-import java.util.concurrent.Callable;
-
-import org.apache.shiro.SecurityUtils;
-import org.apache.shiro.authc.AuthenticationException;
-import org.apache.shiro.authc.AuthenticationToken;
-import org.apache.shiro.authz.AuthorizationException;
-import org.apache.shiro.authz.Permission;
-import org.apache.shiro.session.Session;
-import org.apache.shiro.subject.ExecutionException;
-import org.apache.shiro.subject.PrincipalCollection;
-import org.apache.shiro.subject.Subject;
 
 import com.bakdata.conquery.io.xodus.MasterMetaStorage;
-import com.bakdata.conquery.models.auth.permissions.Ability;
 import com.bakdata.conquery.models.auth.permissions.ConqueryPermission;
-import com.bakdata.conquery.models.auth.util.SinglePrincipalCollection;
 import com.bakdata.conquery.models.exceptions.JSONException;
 import com.bakdata.conquery.models.identifiable.IdentifiableImpl;
 import com.bakdata.conquery.models.identifiable.ids.specific.PermissionOwnerId;
 import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
-import groovy.transform.Synchronized;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
@@ -43,8 +26,7 @@ import lombok.extern.slf4j.Slf4j;
  *            The id type by which an instance is identified
  */
 @Slf4j
-@JsonIgnoreProperties({ "session", "previousPrincipals", "runAs", "principal", "authenticated", "remembered", "principals" })
-public abstract class PermissionOwner<T extends PermissionOwnerId<? extends PermissionOwner<T>>> extends IdentifiableImpl<T> implements Subject {
+public abstract class PermissionOwner<T extends PermissionOwnerId<? extends PermissionOwner<T>>> extends IdentifiableImpl<T>{
 
 	/**
 	 * This getter is only used for the JSON serialization/deserialization
@@ -52,159 +34,9 @@ public abstract class PermissionOwner<T extends PermissionOwnerId<? extends Perm
 	@Getter(value = AccessLevel.PUBLIC, onMethod = @__({@Deprecated}))
 	private final Set<ConqueryPermission> permissions = new HashSet<>();
 
-	@Override
-	public Object getPrincipal() {
-		return getId();
-	}
+	
 
-	@Override
-	public PrincipalCollection getPrincipals() {
-		return new SinglePrincipalCollection(getId());
-	}
-
-	@Override
-	public boolean isPermitted(String permission) {
-		throw new UnsupportedOperationException();
-	}
-
-	@Override
-	public boolean[] isPermitted(String... permissions) {
-		throw new UnsupportedOperationException();
-	}
-
-	@Override
-	public boolean isPermittedAll(String... permissions) {
-		throw new UnsupportedOperationException();
-	}
-
-	@Override
-	public void checkPermission(String permission) throws AuthorizationException {
-		throw new UnsupportedOperationException();
-	}
-
-	@Override
-	public void checkPermissions(String... permissions) throws AuthorizationException {
-		throw new UnsupportedOperationException();
-	}
-
-	@Override
-	public void checkPermission(Permission permission) throws AuthorizationException {
-		if (!(permission instanceof ConqueryPermission)) {
-			throw new AuthorizationException("Supplied permission " + permission + " is not of Type " + ConqueryPermission.class.getName());
-		}
-		SecurityUtils.getSecurityManager().checkPermission(getPrincipals(), permission);
-	}
-
-	@Override
-	public void checkPermissions(Collection<Permission> permissions) throws AuthorizationException {
-		for (Permission permission : permissions) {
-			if (!(permission instanceof ConqueryPermission)) {
-				throw new AuthorizationException(
-					"Supplied permission " + permission + " is not of Type " + ConqueryPermission.class.getName());
-			}
-			SecurityUtils.getSecurityManager().checkPermission(getPrincipals(), permission);
-		}
-	}
-
-	@Override
-	public boolean hasRole(String roleIdentifier) {
-		throw new UnsupportedOperationException();
-	}
-
-	@Override
-	public boolean[] hasRoles(List<String> roleIdentifiers) {
-		throw new UnsupportedOperationException();
-	}
-
-	@Override
-	public boolean hasAllRoles(Collection<String> roleIdentifiers) {
-		throw new UnsupportedOperationException();
-	}
-
-	@Override
-	public void checkRole(String roleIdentifier) throws AuthorizationException {
-		throw new UnsupportedOperationException();
-	}
-
-	@Override
-	public void checkRoles(Collection<String> roleIdentifiers) throws AuthorizationException {
-		throw new UnsupportedOperationException();
-	}
-
-	@Override
-	public void checkRoles(String... roleIdentifiers) throws AuthorizationException {
-		throw new UnsupportedOperationException();
-	}
-
-	@Override
-	public void login(AuthenticationToken token) throws AuthenticationException {
-		throw new UnsupportedOperationException();
-	}
-
-	@Override
-	public Session getSession() {
-		throw new UnsupportedOperationException();
-	}
-
-	@Override
-	public Session getSession(boolean create) {
-		throw new UnsupportedOperationException();
-	}
-
-	@Override
-	public void logout() {
-		throw new UnsupportedOperationException();
-	}
-
-	@Override
-	public <V> V execute(Callable<V> callable) throws ExecutionException {
-		throw new UnsupportedOperationException();
-	}
-
-	@Override
-	public void execute(Runnable runnable) {
-		throw new UnsupportedOperationException();
-	}
-
-	@Override
-	public <V> Callable<V> associateWith(Callable<V> callable) {
-		throw new UnsupportedOperationException();
-	}
-
-	@Override
-	public Runnable associateWith(Runnable runnable) {
-		throw new UnsupportedOperationException();
-	}
-
-	@Override
-	public void runAs(PrincipalCollection principals) throws NullPointerException, IllegalStateException {
-		throw new UnsupportedOperationException();
-	}
-
-	@Override
-	public boolean isRunAs() {
-		throw new UnsupportedOperationException();
-	}
-
-	@Override
-	public PrincipalCollection getPreviousPrincipals() {
-		throw new UnsupportedOperationException();
-	}
-
-	@Override
-	public PrincipalCollection releaseRunAs() {
-		throw new UnsupportedOperationException();
-	}
-
-	@Override
-	public boolean isAuthenticated() {
-		throw new UnsupportedOperationException();
-	}
-
-	@Override
-	public boolean isRemembered() {
-		throw new UnsupportedOperationException();
-	}
+	
 	
 	/**
 	 * Adds permissions to the user and persistent to the storage.
