+++ backend/src/main/java/com/bakdata/conquery/models/auth/subjects/Role.java
@@ -1,14 +1,9 @@
 package com.bakdata.conquery.models.auth.subjects;
 
-import java.util.Collection;
-import java.util.List;
 import java.util.Set;
 
 import javax.validation.constraints.NotNull;
 
-import org.apache.shiro.SecurityUtils;
-import org.apache.shiro.authz.AuthorizationException;
-import org.apache.shiro.authz.Permission;
 import org.hibernate.validator.constraints.NotEmpty;
 
 import com.bakdata.conquery.io.xodus.MasterMetaStorage;
@@ -35,45 +30,7 @@ public class Role extends PermissionOwner<RoleId> {
 	public RoleId createId() {
 		return new RoleId(name);
 	}
-
-	@Override
-	public boolean isPermitted(Permission permission) {
-		return  SecurityUtils.getSecurityManager().isPermitted(getPrincipals(), permission);
-	}
-
-	@Override
-	public boolean[] isPermitted(List<Permission> permissions) {
-		return  SecurityUtils.getSecurityManager().isPermitted(getPrincipals(), permissions);
-	}
-
-	@Override
-	public boolean isPermittedAll(Collection<Permission> permissions) {
-		long numberPermitted = permissions.stream().filter(p -> SecurityUtils.getSecurityManager().isPermitted(getPrincipals(), p)).count();
-		return numberPermitted == permissions.size();
-	}
-
-	@Override
-	public void checkPermission(Permission permission) throws AuthorizationException {
-		throw new UnsupportedOperationException();
-		
-	}
-
-	@Override
-	public void checkPermissions(Collection<Permission> permissions) throws AuthorizationException {
-		throw new UnsupportedOperationException();
-		
-	}
-
-	@Override
-	public boolean isAuthenticated() {
-		throw new UnsupportedOperationException();
-	}
-
-	@Override
-	public boolean isRemembered() {
-		throw new UnsupportedOperationException();
-	}
-
+	
 	@JsonIgnore
 	public Set<ConqueryPermission> getPermissionsEffective() {
 		return getPermissionsCopy();
